requirements.txt
requests
#pymupdf
#pdfplumber
tabula-py
camelot-py[base]
camelot-py[plot]
ghostscript
tk